F12/refactor payment page (#17)
* created PaymentMethod component

* Changed assets

* Added SSL security text

* Small fix
