+++ src/pages/Payment/Payment.js
@@ -4,25 +4,20 @@ import { Redirect } from "react-router-dom";
 import clsx from "clsx";
 import { useFormik } from "formik";
 
-import RadioGroup from "@material-ui/core/RadioGroup";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
 import Button from "../../components/Button";
 import UiInput from "../../components/UiInput";
 import UiInputDate from "../../components/UiInputDate";
-import UiCustomRadio from "../../components/UiCustomRadio";
+import PaymentMethod from "../../components/PaymentMethod";
 
 import paymentSchema from "./Payment-schema";
 
 import withCheckoutLayout from "../../hoc/withCheckoutLayout";
 import CheckoutContext from "../../context/checkout-context";
 
-import paypalImage from "../../img/icons/payment/Paypal.png";
-import applePayImage from "../../img/icons/payment/Apple.png";
-import visaImage from "../../img/icons/payment/Visa.png";
-import mastercardImage from "../../img/icons/payment/MasterCard.png";
-import amexImage from "../../img/icons/payment/AmericanExpress.png";
 import CVV from "../../img/icons/payment/CVV.svg";
+import sslIcon from "../../img/icons/payment/ssl.svg";
 import CreditCard from "../../components/CreditCard";
 
 import { SUMMARY } from "../../constants/routes";
@@ -51,27 +46,6 @@ function Payment() {
     },
   });
 
-  const paymentOptions = [
-    {
-      method: "Credit/Debit Card",
-      formText: "Credit/Debit Card",
-      formImage: false,
-      disabled: false,
-    },
-    {
-      method: "Paypal",
-      formText: false,
-      formImage: paypalImage,
-      disabled: true,
-    },
-    {
-      method: "ApplePay",
-      formText: false,
-      formImage: applePayImage,
-      disabled: true,
-    },
-  ];
-
   function handleFlip() {
     return flip ? setFlip(false) : setFlip(true);
   }
@@ -82,46 +56,10 @@ function Payment() {
           <div className="row">
             <h3>Payment Details</h3>
             <hr />
-            <div className="row">
-              <h5>How would you like to pay?</h5>
-            </div>
-            <div className="row gy-2">
-              <RadioGroup
-                id="paymentMethod"
-                aria-label="payment method"
-                name="paymentMethod"
-                value={formik.values.paymentMethod}
-                onChange={formik.handleChange}
-                row
-              >
-                {paymentOptions.map((payment) => {
-                  return (
-                    <UiCustomRadio
-                      chosenValue={formik.values.paymentMethod}
-                      value={payment.method}
-                      formText={payment.formText}
-                      formImage={payment.formImage}
-                      disabled={payment.disabled}
-                      key={payment.method}
-                    />
-                  );
-                })}
-              </RadioGroup>
-            </div>
-            <div className="row gy-2">
-              <div className="col-12">
-                We accept the following debit/credit cards
-              </div>
-              <div className="col-12">
-                <img className="radio-border-box" src={visaImage} alt="" />
-                <img
-                  className="radio-border-box"
-                  src={mastercardImage}
-                  alt=""
-                />
-                <img className="radio-border-box" src={amexImage} alt="" />
-              </div>
-            </div>
+            <PaymentMethod
+              value={formik.values.paymentMethod}
+              changeHandler={formik.handleChange}
+            />
             <div className="row gy-4">
               <div className="col-6">
                 <UiInput
@@ -212,14 +150,26 @@ function Payment() {
                   </div>
                 </div>
               </div>
-              {/* <div className="col-6">Aquí va la tarjeta</div> */}
-              <CreditCard
-                flip={flip}
-                cardNumber={formik.values.cardNumber}
-                cardCvv={formik.values.cardCvv}
-                cardExpiry={formik.values.cardExpiry}
-                carholderName={formik.values.carholderName}
-              />
+              <div className="col-6">
+                <CreditCard
+                  flip={flip}
+                  cardNumber={formik.values.cardNumber}
+                  cardCvv={formik.values.cardCvv}
+                  cardExpiry={formik.values.cardExpiry}
+                  carholderName={formik.values.carholderName}
+                />
+              </div>
+            </div>
+            <div className="row mt-2">
+              <div className="col-1">
+                <img src={sslIcon} alt="SSL icon" />
+              </div>
+              <div className="col">
+                <p className="mt-3">
+                  We use secure SSL transmission and encrypted storage to
+                  protect your personal information.
+                </p>
+              </div>
             </div>
           </div>
           <div className="row">
